尽量解决unity-warning。由于shader相关不必要warning太多，所以换用debug.log()

+++ TestRayTrace/Assets/HLSL/BVH/BVHCommonDef.hlsl
@@ -28,7 +28,7 @@ HitInfo BVHTraceLocalRay(in Ray ray,
 	tris.GetDimensions(numStructs, stride);
 	int triInxNum = (int)numStructs;
 
-	int start = 0, end = triInxNum / 3 - 1;
+	int start = 0, end = (uint)triInxNum / 3 - 1;
 
 	int toCheck[BVHToCheckMaxNum]; //used as a stack
 	int iter = 0; //used as a stack helper
